third_party/blink/web_tests/wpt_internal/prerender/restriction-media-play.html

<!DOCTYPE html>
<!--
The test aligns with the spec on which behavior is deferred. The test asserts
that media resources are not loaded during prerendering, but it's possible the
spec will allow loading and only disallow playback.
-->
<title>Access to the Play of the Media is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');
  t.add_cleanup(_ => bc.close());

  const types = ['audio', 'video'];
  for (let t = 0; t < types.length; t++) {
    const gotMessage = new Promise(resolve => {
      bc.addEventListener('message', e => {
        resolve(e.data);
      }, {
        once: true
      });
    });

    const url = `resources/media-play.html?type=${types[t]}`;
    window.open(url, '_blank', 'noopener');

    const result = await gotMessage;
    const expected = [
      {
        event: 'started waiting Media.Play',
        prerendering: true
      },
      {
        event: 'prerendering change',
        prerendering: false
      },
      {
        event: 'fired loadedmetadata event after prerendering is activated',
        prerendering: false
      },
      {
        event: 'finished waiting Media.Play',
        prerendering: false
      },
    ];

    assert_equals(result.length, expected.length, `${types[t]}`);
    for (let i = 0; i < result.length; i++) {
      assert_equals(result[i].event, expected[i].event, `event${i}`);
      assert_equals(result[i].prerendering, expected[i].prerendering,
        `prerendering${i}`);
    }
  }
}, `play of the media should be deferred until the prerendered page is
    activated`);

</script>
</body>
